Refactored messages
