+++ exchange/src/exchange/rabbitmq/publisher/RabbitMQPublisher.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "exchange/client_manager/client_manager.hpp"
-#include "shared/messages.hpp"
+#include "shared/messages_exchange_to_wrapper.hpp"
 
 #include <string>
