+++ exchange/src/wrapper/rabbitmq/rabbitmq.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "shared/messages.hpp"
+#include "shared/messages_wrapper_to_exchange.hpp"
+#include "shared/messages_exchange_to_wrapper.hpp"
 #include "wrapper/pywrapper/pywrapper.hpp"
 #include "wrapper/pywrapper/rate_limiter.hpp"
